+++ app/src/main/kotlin/com/thothassistant/stepdaddy/gateway/upstream/CategoryOverrideStore.kt
@@ -24,6 +24,7 @@ object CategoryOverrideStore {
         GroupTitleResolver.DOCUMENTARY,
         GroupTitleResolver.MUSIC,
         GroupTitleResolver.KIDS,
+        GroupTitleResolver.EXTRA_247,
         GroupTitleResolver.INTERNATIONAL,
         GroupTitleResolver.EN_ESPANOL,
         GroupTitleResolver.ADULT,